+++ odbc-api/src/environment.rs
@@ -451,7 +451,7 @@ impl Environment {
                 .environment
                 // Start with first so we are independent of state
                 .drivers_buffer_len(FetchOrientation::First)
-                .into_result(&self.environment)?
+                .into_result_option(&self.environment)?
             {
                 res
             } else {
@@ -463,7 +463,7 @@ impl Environment {
             while let Some((candidate_desc_len, candidate_attr_len)) = self
                 .environment
                 .drivers_buffer_len(FetchOrientation::Next)
-                .into_result(&self.environment)?
+                .into_result_option(&self.environment)?
             {
                 desc_len = max(candidate_desc_len, desc_len);
                 attr_len = max(candidate_attr_len, attr_len);
@@ -575,7 +575,7 @@ impl Environment {
             while let Some((candidate_name_len, candidate_decs_len)) = self
                 .environment
                 .drivers_buffer_len(FetchOrientation::Next)
-                .into_result(&self.environment)?
+                .into_result_option(&self.environment)?
             {
                 server_name_len = max(candidate_name_len, server_name_len);
                 driver_len = max(candidate_decs_len, driver_len);