+++ Changelog.md
@@ -8,6 +8,7 @@
 * `SqlResult::into_result_with_trunaction_check` has been replaced by `SqlResult::into_result_with`.
 * `handles::Environment::driver_connect` now returns `SqlResult<()>` instead of `SqlResult<bool>`.
 * `handles::Environment::drivers_buffer_fill` now returns `SqlResult<()>` instead of `SqlResult<bool>`.
+* `handles::Environment::drivers_buffer_len` now returns `SqlResult<(i16, i16)>` instead of `SqlResult<Option<(i16, i16)>>`.
 * `handles::Statement::fetch` now returns now returns `SqlResult<()>` instead of `SqlResult<bool>`.
 
 ### Non Breaking